+++ NEWS.md
@@ -49,6 +49,8 @@
   files. Fixes #2670.
 * MKVToolNix GUI: multiplexer: the progress dialog shown when scanning a
   Blu-ray wasn't closed properly in certain situations. Fixes #2678.
+* mkvmerge: Matroska reader: mkvmerge did not copy the codec's private data
+  when reading WavPack from Matroska files. Fixes #2685.
 
 
 # Version 40.0.0 "Old Town Road + Pony" 2019-11-09